Added support for both XML and bplist under demoMonkey extension.
NOTE: We don't get correct UTI detection out of the box in this case.  (The reason: OS X determines UTI based on (a) the file extension and (b) the creator code.  DemoMonkey now exports two UTIs with the same extension.  OS X cannot distinguish between the two, since they are both exported from an app with the same creator code.)  This fact informs much of the work in this commit.

- Added DocumentController, a subclass of NSDocumentController, which correctly determines whether a demoMonkey file is an archive (binary plist) or an XML document.  We need this at the level of the document controller so that the de/serialization methods in MyDocument [sic] are called with the correct UTI string.
- Rewrote readFromData:ofType:error: and dataOfType:error: to do the right thing for the the passed-in type.
- Added enum for document type for pretty switching and map from UTI name (typeName) strings to enum values ...for pretty switching.
- AppDelegate instantiates DocumentController on init, making (through that action alone) that instance of DocumentController be the document controller for the application.